+++ packages/koop-logger/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* `LOG_LEVEL` ENV determines whether to log above error level
+
 ## [1.0.1] - 2016-01-08
 ### Fixed
 * Use Winston with built in dailyRotateFile